requirements.txt
psycopg2-binary>=2.9.5
beautifulsoup4>=4.11.1
requests>=2.28.1
anthropic>=0.3.0
python-dotenv>=0.19.0